requirements.txt
streamlit
sqlite3
requests
python-dotenv